unilateral mod-tap: don't send KC_GUI to computer
This fixes a glitch in Miryoku's bilateral-combinations implementation
where the GUI modifiers in a mod-tap key combination are always sent to
the computer alongside the combination, interrupting the user's typing.

For example, suppose you held LGUI_T(KC_A) and then pressed KC_R with
the same hand (assuming your left hand on the standard QWERTY layout).

Before this patch, the combination sends KC_LGUI to the computer when
LGUI_T(KC_A) is held, followed by KC_A and KC_R when finally released.
This inadvertently triggers whatever action is associated with KC_LGUI
on the computer, such as opening the "Start Menu" in Microsoft Windows.

With this patch applied, the combination doesn't send KC_LGUI to the
computer when LGUI_T(KC_A) is held.  Instead, it applies the KC_LGUI
modifier *internally* for bilateral_combinations_hold() and therefore
only sends out KC_A and KC_R to the computer as the user would expect.

However, note that non-GUI modifiers such as Shift/Ctrl/Alt are still
sent to the computer for unilateral mod-tap combinations to allow for
modified mouse clicks such as Shift-click, Ctrl-click, Alt-click, etc.